# Improve JSON column creation performance (list offsets)