## Description
In JSON parser during JSON column creation stage, col_ids is sorted twice - once inside reduce_to_column_tree and before list offset computation.
This PR avoids repeated sort by col_ids for list offset computation. Reuses previously sorted col_ids and node_ids.

Improves parsing bandwidth from 2.170GB/s to 2.414GB/s on GV100 for 1GB json string. (The parsing bandwidth computation includes json string transfer time. Excluding data transfer time, this PR improves parsing bandwidth from 3.9GB/s to 4.5GB/s)

## Checklist
- [x] I am familiar with the [Contributing Guidelines](https://github.com/rapidsai/cudf/blob/HEAD/CONTRIBUTING.md).
- [x] New or existing tests cover these changes.
- [x] The documentation is up to date with these changes.